refactor: findById(), findByAccount() JdbcTemplate 의 queryForObject() 로 구현
