+++ app/src/main/java/com/techcourse/dao/UserDao.java
@@ -1,16 +1,11 @@
 package com.techcourse.dao;
 
-import com.techcourse.dao.Strategy.FindByAccountStrategy;
-import com.techcourse.dao.Strategy.FindByIdStrategy;
 import com.techcourse.domain.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.PreparedStrategy;
 import org.springframework.jdbc.core.RowMapper;
 
-import java.sql.Connection;
-import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.List;
@@ -52,94 +47,12 @@ public class UserDao {
     }
 
     public User findById(final Long id) {
-        Connection conn = null;
-        PreparedStatement pstmt = null;
-        ResultSet rs = null;
-        try {
-            conn = jdbcTemplate.getConnection();
-            PreparedStrategy findByIdStrategy = new FindByIdStrategy();
-            pstmt = findByIdStrategy.createStatement(conn);
-            pstmt.setLong(1, id);
-            rs = pstmt.executeQuery();
-
-            if (rs.next()) {
-                return new User(
-                        rs.getLong(1),
-                        rs.getString(2),
-                        rs.getString(3),
-                        rs.getString(4));
-            }
-            return null;
-        } catch (SQLException e) {
-            log.error(e.getMessage(), e);
-            throw new RuntimeException(e);
-        } finally {
-            try {
-                if (rs != null) {
-                    rs.close();
-                }
-            } catch (SQLException ignored) {
-            }
-
-            try {
-                if (pstmt != null) {
-                    pstmt.close();
-                }
-            } catch (SQLException ignored) {
-            }
-
-            try {
-                if (conn != null) {
-                    conn.close();
-                }
-            } catch (SQLException ignored) {
-            }
-        }
+        return jdbcTemplate.queryForObject("select id, account, password, email from users where id = ?",
+                rowMapper, id);
     }
 
     public User findByAccount(final String account) {
-        Connection conn = null;
-        PreparedStatement pstmt = null;
-        ResultSet rs = null;
-        try {
-            conn = jdbcTemplate.getConnection();
-            PreparedStrategy findByAccountStrategy = new FindByAccountStrategy();
-            pstmt = findByAccountStrategy.createStatement(conn);
-            pstmt.setString(1, account);
-            rs = pstmt.executeQuery();
-
-            if (rs.next()) {
-                return new User(
-                        rs.getLong(1),
-                        rs.getString(2),
-                        rs.getString(3),
-                        rs.getString(4));
-            }
-            return null;
-        } catch (SQLException e) {
-            log.error(e.getMessage(), e);
-            throw new RuntimeException(e);
-        } finally {
-            try {
-                if (rs != null) {
-                    rs.close();
-                }
-            } catch (SQLException ignored) {
-            }
-
-            try {
-                if (pstmt != null) {
-                    pstmt.close();
-                }
-            } catch (SQLException ignored) {
-            }
-
-            try {
-                if (conn != null) {
-                    conn.close();
-                }
-            } catch (SQLException ignored) {
-            }
-        }
+        return jdbcTemplate.queryForObject("select id, account, password, email from users where account = ?"
+                , rowMapper, account);
     }
 }
